// src/app/contenido/interfaces.ts
export type Dioses = "" | "Afrodita" | "Apolo" | "Ares" | "Artemisa" | "Atena" | "Demeter" | "Hefesto" | "Hera" | "Hermes" | "Hestia" | "Poseidon" | "Zeus";